requirements.txt
# Dev/Deployment
pypi-publisher
commonregex
spacy
tableprint
pymysql
pymssql
cx_Oracle
psycopg2-binary
cryptography
peewee
pyyaml
python-magic
pyathena[sqlalchemy]
boto3 == 1.10.34
